refactor(crud): create a file for each type of music instance
