+++ api/quizzify/crud/__init__.py
@@ -1,0 +1,1 @@
+"""CRUD operations for the Quizzify API."""
